Move Segment_2_with_ID to aux
