+++ Straight_skeleton_2/include/CGAL/Straight_skeleton_converter_2.h
@@ -13,7 +13,7 @@
 
 #include <CGAL/license/Straight_skeleton_2.h>
 
-#include <CGAL/Straight_skeleton_2.h>
+#include <CGAL/Straight_skeleton_2/Straight_skeleton_aux.h>
 
 #include <CGAL/assertions.h>
 #include <CGAL/Cartesian_converter.h>
